# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IncidentPilot API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident",
        "operationId": "create_incident_api_v1_incidents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualIncidentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Incident Api V1 Incidents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/analysis": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Start Analysis",
        "operationId": "start_analysis_api_v1_incidents__incident_id__analysis_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Analysis Api V1 Incidents  Incident Id  Analysis Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/timeline": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Timeline",
        "operationId": "list_timeline_api_v1_incidents__incident_id__timeline_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelineEventView"
                  },
                  "title": "Response List Timeline Api V1 Incidents  Incident Id  Timeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/events": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Subscribe Events",
        "operationId": "subscribe_events_api_v1_incidents__incident_id__events_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/evidence": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Evidence",
        "operationId": "list_evidence_api_v1_incidents__incident_id__evidence_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceView"
                  },
                  "title": "Response List Evidence Api V1 Incidents  Incident Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Evidence",
        "operationId": "get_evidence_api_v1_incidents__incident_id__evidence__evidence_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceKind": {
        "type": "string",
        "enum": [
          "metric",
          "log",
          "trace",
          "topology",
          "runbook",
          "change"
        ],
        "title": "EvidenceKind"
      },
      "EvidenceView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "kind": {
            "$ref": "#/components/schemas/EvidenceKind"
          },
          "source_system": {
            "type": "string",
            "title": "Source System"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "query": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query"
          },
          "raw_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Json"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "observed_start": {
            "type": "string",
            "format": "date-time",
            "title": "Observed Start"
          },
          "observed_end": {
            "type": "string",
            "format": "date-time",
            "title": "Observed End"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "collected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Collected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "incident_id",
          "kind",
          "source_system",
          "summary",
          "query",
          "source_uri",
          "observed_start",
          "observed_end",
          "truncated",
          "collected_at"
        ],
        "title": "EvidenceView"
      },
...
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "review",
          "safe_auto",
          "read_only"
        ],
        "title": "ExecutionMode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "IncidentPage"
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "RECEIVED",
          "TRIAGING",
          "INVESTIGATING",
          "SYNTHESIZING",
          "DIAGNOSED",
          "PLANNING",
          "WAITING_APPROVAL",
          "AUTHORIZING",
          "EXECUTING",
          "VERIFYING",
          "ROLLING_BACK",
          "RESOLVED",
          "RESOLVED_READ_ONLY",
          "NEEDS_HUMAN",
          "POLICY_REJECTED",
          "ACTION_FAILED",
          "REJECTED",
          "REPORTING"
        ],
        "title": "IncidentStatus"
      },
      "IncidentView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "source",
          "external_id",
          "status",
          "severity",
          "title",
          "service",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentView"
      },
      "ManualIncidentRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "title": "Description",
            "default": ""
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "service": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Service"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "start_analysis": {
            "type": "boolean",
            "title": "Start Analysis",
            "default": true
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "review"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "severity",
          "service"
        ],
        "title": "ManualIncidentRequest"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "P1",
          "P2",
          "P3",
          "P4"
        ],
        "title": "Severity"
      },
      "TimelineEventView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "actor_type": {
            "type": "string",
            "title": "Actor Type"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "event_type",
          "actor_type",
          "actor_id",
          "payload",
          "created_at"
        ],
        "title": "TimelineEventView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}